requirements.txt
beautifulsoup4==4.10.0
matplotlib==3.5.1
numpy==1.21.4
pandas==1.3.4
Pillow==9.1.0
plotly==5.5.0
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
streamlit==1.3.0
tqdm==4.62.3
